requirements.txt
allure_python_commons==2.9.43
behave==1.2.6
selenium==3.141.0
webdriver_manager==3.5.2
